:sparkles: 테스트 결과 조회 API 구현
테스트 진행 후 결과를 조회하는 API를 구현했습니다.
비회원인 경우 결과 반환이 이루어지고 회원인 경우에는 결과가 저장됩니다.
테스트코드 작성의 편리성을 위해 TestManager로 결과 구현에 필요한 내부 로직을 분리했습니다. 이 내용은 서비스 레이어에서는 비즈니스 로직의 조립 또는 간단한 로직 구현만 이루어짐을 의미합니다.

+++ src/main/java/dnd/donworry/domain/entity/TestResult.java
@@ -5,49 +5,53 @@ import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import lombok.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Entity
 @Builder
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class TestResult extends BaseEntity{
-    @Id
-    @GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)
-    private Long id;
+public class TestResult extends BaseEntity {
+	@Id
+	@GeneratedValue(strategy = jakarta.persistence.GenerationType.IDENTITY)
+	private Long id;
 
-    @Column(nullable = false)
-    private String username;
+	@Column(nullable = false)
+	private String username;
 
-    @Column(nullable = false)
-    private String buddy;
+	@Column(nullable = false)
+	private String buddy;
 
-    @Column(nullable = false)
-    private Long trust;
+	@Column(nullable = false)
+	private Long trust;
 
-    @Column(nullable = false)
-    private Long love;
+	@Column(nullable = false)
+	private Long love;
 
-    @Column(nullable = false)
-    private Long talk;
+	@Column(nullable = false)
+	private Long talk;
 
-    @Column(nullable = false)
-    private Long temperature;
+	@Column(nullable = false)
+	private int temperature;
 
-    @Column(nullable = false)
-    private String imageUrl;
+	@Column(nullable = false)
+	private String imageUrl;
 
-    public static TestResult toEntity (String username, TestResponseDto testResponseDto) {
-        return TestResult.builder()
-                .username(username)
-                .buddy(testResponseDto.getBuddy())
-                .trust(testResponseDto.getTrust())
-                .love(testResponseDto.getLove())
-                .talk(testResponseDto.getTalk())
-                .temperature(testResponseDto.getTemperature())
-                .imageUrl(testResponseDto.getImageUrl())
-                .build();
-    }
+	public static TestResult toEntity(String username, TestResponseDto testResponseDto) {
+		return TestResult.builder()
+			.username(username)
+			.buddy(testResponseDto.getBuddy())
+			.trust(testResponseDto.getTrust())
+			.love(testResponseDto.getLove())
+			.talk(testResponseDto.getTalk())
+			.temperature(testResponseDto.getTemperature())
+			.imageUrl(testResponseDto.getImageUrl())
+			.build();
+	}
 
 }
